Move views for creating group to special folder.
