--- src/StudentToolkit/MVVM/Views/Group/AddStudentsToGroupView.xaml.cs
@@ -1,12 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace StudentToolkit.MVVM.Views.Group
-{
-    public partial class AddStudentsToGroupView : UserControl
-    {
-        public AddStudentsToGroupView()
-        {
-            InitializeComponent();
-        }
-    }
-}
