+++ src/StudentToolkit/MVVM/Views/Group/Create/AddStudentsToGroupView.xaml.cs
@@ -1,0 +1,11 @@
+﻿using System.Windows.Controls;
+
+namespace StudentToolkit.MVVM.Views.Group.Create;
+
+public partial class AddStudentsToGroupView : UserControl
+{
+    public AddStudentsToGroupView()
+    {
+        InitializeComponent();
+    }
+}
