+++ src/StudentToolkit/MVVM/Views/Group/Create/InputGroupDataView.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="StudentToolkit.MVVM.Views.Group.InputGroupDataView"
+﻿<UserControl x:Class="StudentToolkit.MVVM.Views.Group.Create.InputGroupDataView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
@@ -6,7 +6,7 @@
              xmlns:ViewModel="clr-namespace:StudentToolkit.MVVM.ViewModels.Presentation.Group.Create"
              xmlns:CustomControls="clr-namespace:StudentToolkit.WpfCore.CustomControls"
              d:DataContext="{d:DesignInstance Type=ViewModel:InputGroupDataViewModel}"
-             mc:Ignorable="d"
+             mc:Ignorable="d" 
              SizeChanged="OnSizeChanged">
 
     <Grid x:Name="ContentGrid" Margin="200 10">
